Journey test CTCP-1599 covered the user story CTCP-1556, CTCP-1557, CTCP-1619, CTCP-1620, CTCP-1835, Transport folder structure modification
