+++ src/test/scala/uk/gov/hmrc/test/ui/cucumber/stepdefs/Transport/InlandModeOfTransportStepDef.scala
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2022 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.test.ui.cucumber.stepdefs.Transport
+
+import uk.gov.hmrc.test.ui.cucumber.stepdefs.BaseStepDef
+import uk.gov.hmrc.test.ui.pages.Transport.InlandModeOfTransport._
+
+class InlandModeOfTransportStepDef extends BaseStepDef {
+
+  And("""^(?:I )?choose radio option (.*) on the Inland Mode of Transport page$""") { (answer: String) =>
+    InlandModeOfTransportPage
+      .loadPage()
+      .select(answer)
+      .submitPage()
+  }
+
+  And("""^(?:I )?choose radio option (.*) on the Transport Identification page$""") { (answer: String) =>
+    TransportIdentificationPage
+      .loadPage()
+      .select(answer)
+      .submitPage()
+  }
+
+  And("""^(?:I )?enter (.+) as (.+) on the transport identification number page$""") {
+    (answer: String, meansOfTransport: String) =>
+      IdentificationNumberPage
+        .loadPage(meansOfTransport)
+        .fillInput(answer)
+        .submitPage()
+  }
+
+  And("""^(?:I )?select (.+) on the Transport country page$""") { (answer: String) =>
+    TransportCountryPage
+      .loadPage()
+      .select(answer)
+      .submitPage()
+  }
+
+
+}
