+++ src/test/scala/uk/gov/hmrc/test/ui/pages/Transport/InlandModeOfTransport/IdentificationNumberPage.scala
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2022 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.test.ui.pages.Transport.InlandModeOfTransport
+
+import uk.gov.hmrc.test.ui.pages.StringPage
+
+object IdentificationNumberPage extends StringPage {
+
+  override def title(args: String*): String = String.format("What is the %s?", args: _*)
+
+}
